+++ src/tests/integration/components/space-automation/input-stores-form-test.js
@@ -320,37 +320,37 @@ describe('Integration | Component | space automation/input stores form', functio
         });
 
         if (editor === 'rawValue') {
-          const correctInitialValues = complexValuesGenerator(correctValues);
-          const incorrectInitialValues = complexValuesGenerator(incorrectValues);
+          const correctInitialContents = complexValuesGenerator(correctValues);
+          const incorrectInitialContents = complexValuesGenerator(incorrectValues);
 
-          correctInitialValues.forEach(initialValue => {
-            it(`recognizes ${initialValue} value as valid`, async function () {
+          correctInitialContents.forEach(initialContent => {
+            it(`recognizes ${initialContent} value as valid`, async function () {
               const changeSpy = this.get('changeSpy');
               await render(this);
 
-              await fillIn(`.${editor}-field .form-control`, initialValue);
+              await fillIn(`.${editor}-field .form-control`, initialContent);
 
               expect(this.$(`.${editor}-field`)).to.not.have.class('.has-error');
               expect(changeSpy).to.be.calledWith({
                 data: {
-                  s1: JSON.parse(initialValue),
+                  s1: JSON.parse(initialContent),
                 },
                 isValid: true,
               });
             });
           });
 
-          incorrectInitialValues.forEach(initialValue => {
-            it(`recognizes ${initialValue} value as invalid`, async function () {
+          incorrectInitialContents.forEach(initialContent => {
+            it(`recognizes ${initialContent} value as invalid`, async function () {
               const changeSpy = this.get('changeSpy');
               await render(this);
 
-              await fillIn(`.${editor}-field .form-control`, initialValue);
+              await fillIn(`.${editor}-field .form-control`, initialContent);
 
               expect(this.$(`.${editor}-field`)).to.have.class('has-error');
               expect(changeSpy).to.be.calledWith({
                 data: {
-                  s1: JSON.parse(initialValue),
+                  s1: JSON.parse(initialContent),
                 },
                 isValid: false,
               });
@@ -358,7 +358,7 @@ describe('Integration | Component | space automation/input stores form', functio
           });
 
           it('fills initial value with JSON', async function () {
-            const rawValue = JSON.parse(correctInitialValues[0]);
+            const rawValue = JSON.parse(correctInitialContents[0]);
             getStores(this)[0].defaultInitialContent = rawValue;
 
             await render(this);